Added some more project get tests
